+++ Plugins/CoreModel/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/coremodel/PluginCoreModel.java
@@ -54,7 +54,6 @@ import de.fraunhofer.iosb.ilt.frostserver.util.exception.UpgradeFailedException;
 import java.io.IOException;
 import java.io.Writer;
 import java.util.Arrays;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -371,40 +370,40 @@ public class PluginCoreModel implements PluginRootDocument, PluginModel, Liquiba
         return true;
     }
 
-    public Map<String, Object> createLiqibaseParams(JooqPersistenceManager ppm, Map<String, Object> target) {
-        if (target == null) {
-            target = new LinkedHashMap<>();
+    @Override
+    public Map<String, Object> createLiqibaseParams(PersistenceManager pm, Map<String, Object> target) {
+        if (pm instanceof JooqPersistenceManager ppm) {
+            ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_DATASTREAM, modelSettings.idTypeDatastream);
+            ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_FEATURE, modelSettings.idTypeFeature);
+            ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_HIST_LOCATION, modelSettings.idTypeHistLoc);
+            ppm.generateLiquibaseVariables(target, NAME_ET_HISTORICALLOCATION, modelSettings.idTypeHistLoc);
+            ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_LOCATION, modelSettings.idTypeLocation);
+            ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_OBS_PROP, modelSettings.idTypeObsProp);
+            ppm.generateLiquibaseVariables(target, NAME_ET_OBSERVEDPROPERTY, modelSettings.idTypeObsProp);
+            ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_OBSERVATION, modelSettings.idTypeObservation);
+            ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_SENSOR, modelSettings.idTypeSensor);
+            ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_THING, modelSettings.idTypeThing);
         }
-        target.put(CHANGE_SET_NAME, "Plugin.CoreModel");
-        ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_DATASTREAM, modelSettings.idTypeDatastream);
-        ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_FEATURE, modelSettings.idTypeFeature);
-        ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_HIST_LOCATION, modelSettings.idTypeHistLoc);
-        ppm.generateLiquibaseVariables(target, NAME_ET_HISTORICALLOCATION, modelSettings.idTypeHistLoc);
-        ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_LOCATION, modelSettings.idTypeLocation);
-        ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_OBS_PROP, modelSettings.idTypeObsProp);
-        ppm.generateLiquibaseVariables(target, NAME_ET_OBSERVEDPROPERTY, modelSettings.idTypeObsProp);
-        ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_OBSERVATION, modelSettings.idTypeObservation);
-        ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_SENSOR, modelSettings.idTypeSensor);
-        ppm.generateLiquibaseVariables(target, NAME_LIQUIBASE_THING, modelSettings.idTypeThing);
-
         return target;
     }
 
     @Override
-    public String checkForUpgrades() {
+    public String checkForUpgrades(Map<String, Object> liquibaseParams) {
+        liquibaseParams.put(CHANGE_SET_NAME, "Plugin.CoreModel");
         try (PersistenceManager pm = PersistenceManagerFactory.getInstance(settings).create()) {
             if (pm instanceof JooqPersistenceManager jpm) {
-                return jpm.checkForUpgrades(LIQUIBASE_CHANGELOG_FILENAME, createLiqibaseParams(jpm, null));
+                return jpm.checkForUpgrades(LIQUIBASE_CHANGELOG_FILENAME, liquibaseParams);
             }
             return "Unknown persistence manager class";
         }
     }
 
     @Override
-    public boolean doUpgrades(Writer out) throws UpgradeFailedException, IOException {
+    public boolean doUpgrades(Writer out, Map<String, Object> liquibaseParams) throws UpgradeFailedException, IOException {
+        liquibaseParams.put(CHANGE_SET_NAME, "Plugin.CoreModel");
         try (PersistenceManager pm = PersistenceManagerFactory.getInstance(settings).create()) {
             if (pm instanceof JooqPersistenceManager jpm) {
-                return jpm.doUpgrades(LIQUIBASE_CHANGELOG_FILENAME, createLiqibaseParams(jpm, null), out);
+                return jpm.doUpgrades(LIQUIBASE_CHANGELOG_FILENAME, liquibaseParams, out);
             }
             out.append("Unknown persistence manager class");
             return false;